requirements.txt
fastapi
uvicorn
gitpython
pinecone-client
sentence-transformers
langchain
langchain-community
openai
python-dotenv